Project file to be added here.
The MainActivity.java is the runnable, however it needs to be put in the src\main\java subdirectory.  This project was created using JSONArrays and JSONObjects.  Upon running this java program, you will get prompted to enter a location, and then proceeding will have the program result in the weather for that location using an API.